+++ keyring/core.py
@@ -6,6 +6,7 @@ Created by Kang Zhang on 2009-07-09
 import os
 import sys
 import warnings
+import logging
 
 from .py27compat import configparser
 
@@ -13,6 +14,7 @@ from keyring import logger
 from keyring import backend
 from keyring.util import platform_ as platform
 
+log = logging.getLogger(__name__)
 
 _keyring_backend = None
 
@@ -53,6 +55,7 @@ def init_backend():
     """
     Load a keyring specified in the config file or infer the best available.
     """
+    _load_library_extensions()
     set_keyring(load_config() or _get_best_keyring())
 
 
@@ -138,6 +141,40 @@ def _load_keyring_path(config):
         sys.path.insert(0, path)
     except (configparser.NoOptionError, configparser.NoSectionError):
         pass
+
+def _load_library_extensions():
+    """
+    Locate all setuptools entry points by the name 'keyring backends'
+    and initialize them.
+
+    Any third-party library may register an entry point by adding the
+    following to their setup.py::
+
+        entry_points = {
+            'keyring backends': [
+                'plugin name = mylib.mymodule:initialize_func',
+            ],
+        },
+
+    `plugin name` can be anything.
+    `initialize_func` is optional and will be invoked by keyring on startup.
+
+    Most plugins will simply provide or import a KeyringBackend in `mymodule`.
+    """
+    group = 'keyring backends'
+    try:
+        pkg_resources = __import__('pkg_resources')
+    except ImportError:
+        return
+    entry_points = pkg_resources.iter_entry_points(group=group)
+    for ep in entry_points:
+        try:
+            log.info('Loading keyring backends from %s', ep.name)
+            init_func = ep.load()
+            if callable(init_func):
+                init_func()
+        except Exception as exc:
+            log.exception("Error initializing plugin %s (%s).", ep, exc)
 
 # init the _keyring_backend
 init_backend()